% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_predict_tRackIT.R
\name{activity.predict.tRackIT}
\alias{activity.predict.tRackIT}
\title{activity classification}
\usage{
activity.predict.tRackIT(animal)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}
}
\description{
classifies data based on ML variables
}
\author{
Jannis Gottwald
}
