% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_vars_tRackIT.R
\name{activity.vars.tRackIT}
\alias{activity.vars.tRackIT}
\title{calculate variables for activity classification}
\usage{
activity.vars.tRackIT(animal, tcol, scol, dcol, tzone, rscale = 0)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{tcol}{string, colname of timestamp column}

\item{scol}{string, colname of signal strength column}

\item{dcol}{string, colname of signal duration column}

\item{tzone}{string, time zone of timestamp}

\item{rscale}{numeric, numericvalue to rescale to dBW scale}
}
\description{
calculates variables for the ML-based activity classification
}
\author{
Jannis Gottwald
}
