% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_aggregate_tRackIT.R
\name{activity_aggregate_tRackIT}
\alias{activity_aggregate_tRackIT}
\title{activity aggregatio}
\usage{
activity_aggregate_tRackIT(animal, avalue, tzone)
}
\arguments{
\item{animal}{list, list generated by initAnimal function}

\item{avalue}{numeric, aggregation interval in minutes}

\item{tzone}{string, Timezone of time data}
}
\description{
aggregates the activity classification to user specified intervalls of "avalue" minutes by using the class with max entries (mode) in interval
}
\examples{
anml <- getAnimal(projList = test_project, projroot = "test_project/", animalID = "woodpecker")
activity_aggregate_tRackIT(animal = anml, avalue = 1, tzone = "CET")

}
\author{
Jannis Gottwald
}
