% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_predict_tRackIT.R
\name{activity_predict_tRackIT}
\alias{activity_predict_tRackIT}
\title{activity classification}
\usage{
activity_predict_tRackIT(animal)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}
}
\description{
classifies data based on ML variables
}
\examples{
# get animal
anml <- getAnimal(projList = test_project, projroot = "/test_project/", animalID = "woodpecker")
# predict
activity_predict_tRackIT(animal = anml)

}
\author{
Jannis Gottwald
}
