% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_vars_tRackIT.R
\name{activity_vars_tRackIT}
\alias{activity_vars_tRackIT}
\title{calculate variables for activity classification}
\usage{
activity_vars_tRackIT(animal, tcol, scol, dcol, tzone, rscale = 0)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{tcol}{string, colname of timestamp column}

\item{scol}{string, colname of signal strength column}

\item{dcol}{string, colname of signal duration column}

\item{tzone}{string, time zone of timestamp}

\item{rscale}{numeric, numericvalue to rescale to dBW scale}
}
\description{
calculates variables for the ML-based activity classification
}
\examples{
# get animal
anml <- getAnimal(projList = test_project, projroot = "/test_project/", animalID = "woodpecker")
# calculate activity vars
activity_vars_tRackIT(animal = anml, tcol = "timestamp", scol = "max", dcol = "receiver", tzone = "CET", rscale = 0)

}
\author{
Jannis Gottwald
}
