% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_add_ID_info.R
\name{add.ID.Info}
\alias{add.ID.Info}
\title{adds meta data}
\usage{
add.ID.Info(animal, data)
}
\arguments{
\item{animal}{list, generated by initAnimal function}

\item{data}{data.frame, data frame data.table}
}
\description{
adds meta data from idFile.rds to data
}
\author{
Jannis Gottwald
}
