% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_awk_tRackIT.R
\name{batch.awk.tRackIT}
\alias{batch.awk.tRackIT}
\title{create batch file to prefilter tagged individuals using awk, Installation following the instruction (http://gnuwin32.sourceforge.net/packages/gawk.htm) necessary. Execution of multiple.bat files possible}
\usage{
batch.awk.tRackIT(
  animal,
  projList,
  freq_er_l,
  freq_er_r,
  min_dur,
  max_dur,
  in_project,
  path_to_folder,
  OS
)
}
\arguments{
\item{animal}{list, generated by initAnimal function}

\item{projList}{list, generatet by initProject function}

\item{freq_er_l}{num, accepted frequency range minus tag frequency (khz)}

\item{freq_er_r}{num, accepted frequency  range plus tag frequency (khz)}

\item{min_dur}{num, minimum duration of signal}

\item{max_dur}{num, maximum duration of signal}

\item{in_project}{logical, if true the awk command is created from the data in the project folder}

\item{path_to_folder}{string, awk command is created from the data in the indicated folder}

\item{OS}{string, operating system ("Windows", "MacOS", "Linux")}
}
\description{
Creates .bat files for prefiltering with awk (only for data recorded with tRackIT.OS)
}
\author{
Jannis Gottwald
}
