% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bearings_tRackIT.R
\name{calc.bearings.tRackIT}
\alias{calc.bearings.tRackIT}
\title{bearing calculation}
\usage{
calc.bearings.tRackIT(path_to_data, dbLoss, animal, projList)
}
\arguments{
\item{path_to_data}{string, path to filtered file}

\item{dbLoss}{num, gain loss from 0 to 90° antenna orientation}

\item{animal}{list, generatet by initanimal}
}
\description{
calculates bearings using linear approximation and acos
}
\author{
Jannis Gottwald
}
