% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bearings_tRackIT.R
\name{calc_bearings_tRackIT}
\alias{calc_bearings_tRackIT}
\title{bearing calculation}
\usage{
calc_bearings_tRackIT(path_to_data, dbLoss, animal, projList)
}
\arguments{
\item{path_to_data}{string, path to filtered file}

\item{dbLoss}{num, gain loss from 0 to 90° antenna orientation}

\item{animal}{list, generatet by initAnimal()}

\item{projList}{list, generatet by initProject()}
}
\description{
calculates bearings using linear approximation and acos. 0,90.180,270 degrees Antenna orinetations are currently expected
}
\examples{

projroot<-paste0(getwd(),"/tRackIT_test_data/")
p <- getProject(projroot)
animal <- getAnimal(projroot =projroot, animalID = "woodpecker")
m_logger <- list.files(animal$path$logger_timematch, full.names = TRUE)
lapply(m_logger, function(x) calc_bearings_tRackIT(x, dbLoss = 21, animal = animal, projList = p))

}
\author{
Jannis Gottwald
}
