% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_triangulations_tRackIT.R
\name{calc_telemetr_tRackIT}
\alias{calc_telemetr_tRackIT}
\title{calculate triangulations using telemetr}
\usage{
calc_telemetr_tRackIT(data, tmstmp)
}
\arguments{
\item{data}{spatial.data.frame,}

\item{tmstmp}{character, central timestamp of rolling window}
}
\description{
calculates triangulations using functionalities from telemetr-package
}
\author{
Jannis Gottwald
}
