% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{Calibration values}
\usage{
calibrate(projList, path_to_data)
}
\arguments{
\item{projList}{list, list generated by initProject function}

\item{path_to_data}{string, path to filtered file}
}
\description{
Get correction values for signal strength from calibration curves
}
\author{
RadioTracking EU
}
