% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_catalogue.R
\name{data.catalogue}
\alias{data.catalogue}
\title{catalogue of all tRackIT csv with start and end date}
\usage{
data.catalogue(projList, in_project = TRUE, path_to_folder = ".")
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{in_project}{logical, if true the catalogue is created from the data in the project folder}

\item{path_to_folder}{string, e catalogue is created from the data in the indicated folder}
}
\description{
Creates a catalogue of all raw logger files in folder
}
\author{
Jannis Gottwald
}
