% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_stat.R
\name{dist_to_stat}
\alias{dist_to_stat}
\title{distance to station}
\usage{
dist_to_stat(data, projList)
}
\arguments{
\item{data}{data.frame, generatet by triangulation function}

\item{antennas}{file, data.frame with antenna information (Lat,Lon,orientation)}
}
\description{
calculates the distance of traingulations to the station with maxDB
}
\author{
Jannis Gottwald
}
