% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_catalogue.R
\name{file.catalogue}
\alias{file.catalogue}
\title{create file and parameter catalogue from raw logger data}
\usage{
file.catalogue(projList, station = ".", collection = ".")
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{station}{string, station of hich the files should be catagolized}

\item{collection}{string, path to data collection in station folder from which the files should be catagolized}
}
\description{
Creates a catalogue of all raw logger files in folder
}
\author{
Jannis Gottwald
}
