% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tRackIT.R
\name{filter.tRackIT}
\alias{filter.tRackIT}
\title{filter automatic (tRackIT.OS)}
\usage{
filter.tRackIT(projList, animal, freq_er = 2, path_to_data, d_min, d_max)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{animal}{list, generated by initanimal function}

\item{freq_er}{numeric, accepted deviation from tag frequency}

\item{path_to_data}{string, path to awk filtered file}

\item{d_min}{numeric, minimum duration of signal}

\item{d_max}{numeric, maximum duration of signal}
}
\description{
automatic filtering of awk filtered data (only based on tRackIT.OS). Awk files may miss colnames. make shure they have colnames(data)="timestamp", "samples", "duration", "signal_freq", "signal_bw", "max_signal","noise", "receiver", "station"
}
\author{
Jannis Gottwald
}
