% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_triangulations.R
\name{filter.triangs.tRackIT}
\alias{filter.triangs.tRackIT}
\title{filter triangulations}
\usage{
filter.triangs.tRackIT(
  animal,
  projList,
  rerange = 400,
  nbi = 1,
  speed = 20,
  tz = "UTC",
  filter_speed = TRUE,
  filter_distance = TRUE,
  filter_biang = TRUE,
  tw = 5
)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{projList}{list, list generated by initProject functiom}

\item{rerange}{numeric, max accepted reception range of single stations, points that are farer than rerange from loudest station will be filtered out}

\item{nbi}{numeric, number of minimum succesul biangulations (intersections) in group}

\item{speed}{numeric, maximum accepted speed between consecutive points}

\item{tz}{string, timezone of data time}

\item{filter_speed}{logical, if TRUE speed filter will be applied}

\item{filter_distance, }{if TRUE distance filter will be applied}

\item{filter_biang, }{if TRUE biang filter will be applied}

\item{tw}{numeric, time window of triangulations}
}
\description{
filters triangulations based on different filter options
}
\author{
Jannis Gottwald
}
