% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tRackIT.R
\name{filter_tRackIT}
\alias{filter_tRackIT}
\title{filter automatic (tRackIT.OS)}
\usage{
filter_tRackIT(
  animal = NULL,
  freq_er = 2,
  path_to_data = NULL,
  d_min = NULL,
  d_max = NULL,
  start = NULL,
  end = NULL,
  freq = NULL
)
}
\arguments{
\item{animal}{list, generated by initanimal function}

\item{freq_er}{numeric, accepted deviation from tag frequency in khz}

\item{path_to_data}{string, path to awk filtered file}

\item{d_min}{numeric, minimum duration of signal}

\item{d_max}{numeric, maximum duration of signal}

\item{start}{string, start of tracking YYYY-MM-DD}

\item{end}{string, end of tracking YYYY-MM-DD}

\item{freq}{num, tag frequency (khz)}
}
\description{
Filtering of raw signals based on transmitter frequency, signal duration and runtime of the transmitter. The function reads files from a given folder. Make shure that there is one file per station and that they have the following columns:timestamp= time of received signal-expected format: “\%Y-\%m-\%d \%H:\%M:\%S (e.g. 1970-01-29 15:30:22:11),  duration=signal length of the recorded signal, frequency= frequency in kHz (e.g. 150150.2),  max= the max signal strength in dBW of the recorded signal (peak) , receiver= name of the receiver device (usually 0,1,2,3 for a 4 Antenna station ), station= name of the station as character string
}
\examples{
#projroot<-paste0(getwd(),"/tRackIT_test_data/")
#anml<-getAnimal(projroot =projroot, animalID = "woodpecker")
#filter_tRackIT(animal = anml, freq_er = 2, path_to_data = "/test_project/data/logger_data_csv/", d_min = anml$meta$duration_min, d_max = anml$meta$duration_max, freq = anml$meta$freq, start = anml$meta$start, end = anml$meta$end)

}
\author{
Jannis Gottwald
}
