% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{get radio-tracking project}
\usage{
getProject(projroot = ".", plot = FALSE)
}
\arguments{
\item{projroot}{string, project directory}

\item{plot}{logical, if TRUE a map with stations will be plottet}
}
\description{
reads project file from projroot
}
\examples{
# get project
projroot<-paste0(getwd(),"/tRackIT_test_data/")
test_project <- getProject(projroot = projroot, plot = FALSE)

}
\author{
Jannis Gottwald
}
