\name{gmt3}
\alias{gmt3}
\docType{data}
\title{gmt3}
\description{
  A dataset of land polygons used for plotting maps
}
\usage{data(gmt3)}
\format{
  A data frame with 504522 observations on the following 2 variables.
  \describe{
    \item{\code{longitude}}{a numeric vector}
    \item{\code{latitude}}{a numeric vector}
  }
}
\source{
  \url{http://gmt.soest.hawaii.edu/}
}
\examples{
  data(gmt3)
}
\keyword{datasets}
