% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel_time.R
\name{hampel_time}
\alias{hampel_time}
\title{hampel filter}
\usage{
hampel_time(path_to_data, col, k, t0 = 0.5, animal)
}
\arguments{
\item{path_to_data}{string, dpath to file containing bearings}

\item{col}{string, string indicating the column containing bearings}

\item{k}{numeric, window size n seconds}

\item{t0}{numeric,threshold for median filter. A high threshold makes the filter more forgiving, a low one will declare more points to be outliers. Default=0.5}
}
\description{
detect outliers in bearings using hampel filter in a moving time window
}
\examples{
projroot <- "/test_project/"
animal <- getAnimal(projroot, animalID = "woodpecker")
bearings <- list.files(animal$path$bearings, full.names = TRUE)
lapply(bearings, function(x) hampel_time(path_to_data = x, col = "linear", animal = animal, k = 10))

}
\author{
Jannis Gottwald
}
