% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initAnimal.R
\name{initAnimal}
\alias{initAnimal}
\title{Init Animal Tracking}
\usage{
initAnimal(
  projList = NULL,
  projroot = ".",
  saveAnml = FALSE,
  animalID = NULL,
  species = NA,
  sex = NA,
  age = NA,
  weight = NA,
  rep.state = NA,
  freq = NA,
  start = NA,
  end = NA,
  duration_min = NA,
  duration_max = NA
)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{projroot}{string, project directory}

\item{animalID}{string, label of the tagged individual}

\item{species}{string, species of the tagged individual}

\item{sex}{string, sex of the tagged individual}

\item{age}{string, age of the tagged individual}

\item{weight}{string, weight of the tagged individual}

\item{rep.state}{string, reproductive state of the tagged individual}

\item{freq}{num, tag frequency (khz)}

\item{start}{string, start of tracking YYYY-MM-DD}

\item{end}{string, end of tracking YYYY-MM-DD}

\item{duration_min}{numeric, minimum duration of single signal}

\item{duration_max}{numeric, maximum duration of single signal}

\item{logical, }{if TRUE animal file will be stored}
}
\description{
Creates directories and a R list with all the needed tRackIT information for one animal. R list is stored as .rds in animals directory-projroot/data/individuals/animalID/animalID_idFile.rds
}
\examples{
# get project file
#projroot<-paste0(getwd(),"/tRackIT_test_data/")
#test_project <- getProject(projroot = projroot, plot = TRUE)
# init animal
#initAnimal(projList = test_project, projroot = "test_project/", saveAnml = TRUE, animalID = test_project$tags$ID[1], species = "woodpecker", sex = "m", age = "adult", weight = 36, rep.state = "breeding", freq = test_project$tags$frequency[1], start = test_project$tags$start[1], end = test_project$tags$end[1], duration_min = 0.012, duration_max = 0.4)

}
\author{
Marvin Ludwig & Jannis Gottwald
}
