% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProject.R
\name{initProject}
\alias{initProject}
\title{Init radio-tracking project}
\usage{
initProject(
  projroot = ".",
  logger_data_raw,
  tags = NULL,
  id_col = NULL,
  start_col = NULL,
  end_col = NULL,
  freq_col = NULL,
  dmin_col = NULL,
  dmax_col = NULL,
  stations = NULL,
  s_col,
  x_col,
  y_col,
  r_col,
  o_col,
  epsg = NULL,
  tz = NULL
)
}
\arguments{
\item{projroot}{string, project directory}

\item{logger_data_raw}{string, path to the full RTS dataset}

\item{tags}{data.frame, table containing all relevant information of tagged individuals}

\item{id_col}{string, name of the column in tags data.frame containing the name (ID) of the tagged individual}

\item{start_col}{string, name of the column in tags data.frame containing the date of the tagging (yy-mm-dd)}

\item{end_col}{string, name of the column in tags data.frame containing the end of the transmitter lifetime (yy-mm-dd)}

\item{freq_col}{string, name of the column in tags data.frame containing the frequency of the transmitter in kHz}

\item{dmin_col}{string, name of the column in tags data.frame containing the min duration of the transmitter signal in seconds (10ms=0.01 sec)}

\item{dmax_col}{string, name of the column in tags data.frame containing the max duration of the transmitter signal in seconds (10ms=0.01 sec)}

\item{stations}{data.frame, with station coordinates, station name, receiver name and antenna orientation}

\item{s_col}{string, name of the column in tags data.frame containing the name of each station}

\item{x_col}{string, name of the column in tags data.frame containing the X Coordinates}

\item{y_col}{string, name of the column in tags data.frame containing the Y Coordinates}

\item{r_col}{string, name of the column in tags data.frame containing the name of each receiver}

\item{o_col}{string, name of the column in tags data.frame containing the orientation of antennas}

\item{epsg}{numeric, epsg code of the coordinate system of stations coordinates- will be transformed to latlon}

\item{tz}{string, timezone of project}
}
\description{
Creates directories and a R list with all the needed tRackIT-Project information. R list is stored as .rds in projroot directory-projroot/projID_idFile.rds
}
\examples{

projroot<-paste0(getwd(),"/tRackIT_test_data/")
# data frame with id information
tag_df <- data.frame(id = "woodpecker", freq = 150050, start = "2021-06-10", end = "2021-06-14", dmax = 0.025, dmin = 0.012)
# data frame with tRackIt station information
stations <- data.frame(station = "station1", receiver = c(0, 1, 2, 3), orientation = c(0, 90, 180, 270), X = 8.677942, Y = 50.84420)
test_project <- initProject(projroot = projroot, logger_data_raw = ".", tags = tag_df, id_col = "id", start_col = "start", end_col = "end", freq_col = "freq", dmin_col = "dmin", dmax_col = "dmax", stations = stations, s_col = "station", x_col = "X", y_col = "Y", r_col = "receiver", o_col = "orientation", epsg = 4326, tz = "CET")
tst$path

}
\author{
Jannis Gottwald
}
