\name{print.trackit}
\alias{print.trackit}
\title{Print trackit object}
\description{
  Prints a pretty summary of an object of class \code{trackit}
}
\usage{
print.trackit(x, ...)
}
\arguments{
  \item{x}{an object of class \code{trackit} typically generated
           with the \code{\link{trackit}} function.}
  \item{\dots}{additional arguments}
}
\author{Anders Nielsen \email{anders.nielsen@hawaii.edu}, John Sibert \email{sibert@hawaii.edu}}
\seealso{\code{\link{trackit}}}
\keyword{models}

