% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_logger_data_tRackIT.R
\name{read.logger.data.tRackIT}
\alias{read.logger.data.tRackIT}
\title{join raw logger data per station}
\usage{
read.logger.data.tRackIT(projList)
}
\arguments{
\item{projList}{list, generatet by initProject function}
}
\description{
joins all raw logger files in a folder (only for files recorded with tRackit.OS!)
}
\author{
Jannis Gottwald
}
