% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtRoot.R
\name{rtRoot}
\alias{rtRoot}
\title{Initialise a Radiotracking Project}
\usage{
rtRoot(projroot = ".")
}
\arguments{
\item{projroot}{string, path to home directory of the project}
}
\description{
Initialise a Radiotracking Project
}
