\name{scroll}
\alias{scroll}
\title{Scrolls through the light record}
\description{Scrolls through the light record}
\usage{
scroll(track, n=1000)
}
\arguments{
  \item{track}{A data.frame with a light column}
  \item{n}{An integer step size}
}
\value{
  No value is returned. This function is invoked for its side effects.
}
\author{Anders Nielsen \email{anders.nielsen@hawaii.edu} and John Sibert \email{sibert@hawaii.edu}}

\keyword{models}
