% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match_logger_tRackIT.R
\name{time_match_logger_tRackIT}
\alias{time_match_logger_tRackIT}
\title{time match logger data}
\usage{
time_match_logger_tRackIT(animal = NULL, path_to_data)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{path_to_data}{string, path to filtered data}
}
\description{
timematch of simoultaneously received signals by logger of one station
}
\examples{
#projroot<-paste0(getwd(),"/tRackIT_test_data/")

#p <- getProject(projroot)
#animal <- getAnimal(projroot, animalID = "woodpecker")
#fls <- list.files(animal$path$filtered, full.names = TRUE)
#lapply(fls, function(x) time_match_logger_tRackIT(animal, x))

}
\author{
Jannis Gottwald
}
