% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match_station.R
\name{time_match_station}
\alias{time_match_station}
\title{time match logger data}
\usage{
time_match_station(data, method, tmstmp)
}
\arguments{
\item{data}{data.frame, data containing vhf-signals}

\item{method}{string, which bearing calculation method ("linear", "acos", "ML")}

\item{tmstmp}{string, timestamp}
}
\description{
timematch of simoultaneously received signals by stations
}
\author{
Jannis Gottwald
}
