% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day.R
\name{time_of_day}
\alias{time_of_day}
\title{calculates time of day information}
\usage{
time_of_day(data, Lat, Lon, tcol, tz, activity_period)
}
\arguments{
\item{data}{data.frame, data with timestamps}

\item{Lat}{numeric, reference latitude for sunrise/sunset calculation}

\item{Lon}{numeric, reference longitude for sunrise/sunset calculation}

\item{tcol}{string, colname of timestamp column}

\item{tz}{string, time zone of region ton convert timezone of data to the right format}

\item{activity_period}{string, either "diurnal" or "nocturnal"}
}
\description{
calculates sunset and sunrise for each data point and the distance to it. For nocturnal species sunrise is +1 day of sunset
}
\author{
Jannis Gottwald
}
