\name{trackit-internal}
\title{Internal trackit objects}
\alias{approxDeltaT} 
\alias{deg.to.rad} 
\alias{JD} 
\alias{JDE} 
\alias{obliquity} 
\alias{rad.to.deg} 
\alias{sidereal.UT}
\alias{solar.altitude} 
\alias{solar.coordinates} 
\alias{solar.coordinates.local}
\description{Internal trackit objects.}
\details{These are not to be called by the user.}
\keyword{internal}
 
