% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{filter data interactive}
\usage{
triangulate(method, path_to_data, animal, projList, collision = FALSE)
}
\arguments{
\item{method}{string,("linear, "acos", "ML") indicates wether linear, acos ore machine learning based bearings are used}

\item{path_to_data}{string, path to filtered file}

\item{animal}{list, generatet by initanimal}

\item{projList}{list, generatet by initProject}

\item{collision}{logical, if collision corrected data is used a suffix is added to resulting file name}
}
\description{
calculate triangulations based on different bearing calculation methods
}
\author{
Jannis Gottwald
}
