% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_aggregate_tRackIT.R
\name{activity_aggregate_tRackIT}
\alias{activity_aggregate_tRackIT}
\title{activity aggregatio}
\usage{
activity_aggregate_tRackIT(animal, avalue, tzone, get_data = FALSE)
}
\arguments{
\item{animal}{list, list generated by initAnimal function}

\item{avalue}{numeric, aggregation interval in minutes}

\item{tzone}{string, Timezone of time data}

\item{get_data}{logical, if TRUE classification will be returned as data.frame and stored in the classification folder. If FALSE classification will only be stored}
}
\description{
aggregates the activity classification to user specified intervalls of "avalue" minutes by using the class with max entries (mode) in interval
}
\examples{
projroot<-paste0(getwd(),"/tRackIT_test_data/")
anml<-getAnimal(projroot =projroot, animalID = "woodpecker")
activity_aggregate_tRackIT(animal = anml, avalue = 1, tzone = "CET")

}
\author{
Jannis Gottwald
}
