% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_vars_tRackIT.R
\name{activity_vars_tRackIT}
\alias{activity_vars_tRackIT}
\title{calculate variables for activity classification}
\usage{
activity_vars_tRackIT(animal, t_col, s_col, r_col, tz, rscale = 0)
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{t_col}{string, colname of timestamp column}

\item{s_col}{string, colname of signal strength column}

\item{r_col}{tring, name of the column containing the name of each receiver}

\item{tz}{string, time zone of timestamp}

\item{rscale}{numeric, numericvalue to rescale to dBW scale}
}
\description{
calculates variables for the ML-based activity classification
}
\examples{
# get animal
#projroot<-paste0(getwd(),"/tRackIT_test_data/")
#anml<-getAnimal(projroot =projroot, animalID = "woodpecker")
# calculate activity vars
#activity_vars_tRackIT(animal = anml, t_col = "timestamp", s_col = "max", r_col = "receiver", tz = "CET", rscale = 0)

}
\author{
Jannis Gottwald
}
