% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimal.R
\name{getAnimal}
\alias{getAnimal}
\title{get Animal}
\usage{
getAnimal(projList = ".", animalID)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{animalID}{string, label of the tagged individual}
}
\description{
reads anmimal file from animal root
}
\author{
Jannis Gottwald
}
