% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimal.R
\name{getAnimal}
\alias{getAnimal}
\title{get Animal}
\usage{
getAnimal(projroot = ".", animalID)
}
\arguments{
\item{projroot}{string, project directory}

\item{animalID}{string, label of the tagged individual}
}
\description{
reads anmimal file from animal root
}
\examples{
projroot<-paste0(getwd(),"/tRackIT_test_data/")
getAnimal(projroot = projroot, animalID = "woodpecker")

}
\author{
Jannis Gottwald
}
