% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initAnimal.R
\name{initAnimal}
\alias{initAnimal}
\title{Init Animal Tracking}
\usage{
initAnimal(
  projList,
  saveAnml = FALSE,
  animalID,
  species = NA,
  sex = NA,
  age = NA,
  weight = NA,
  rep.state = NA,
  freq = NA,
  start = NA,
  end = NA,
  duration_min = NA,
  duration_max = NA
)
}
\arguments{
\item{projList}{list, generatet by initProject function}

\item{animalID}{string, label of the tagged individual}

\item{species}{string, species of the tagged individual}

\item{sex}{string, sex of the tagged individual}

\item{age}{string, age of the tagged individual}

\item{weight}{string, weight of the tagged individual}

\item{rep.state}{string, reproductive state of the tagged individual}

\item{freq}{num, tag frequency (khz)}

\item{start}{string, start of tracking YYYY-MM-DD}

\item{end}{string, end of tracking YYYY-MM-DD}

\item{duration_min}{numeric, minimum duration of single signal}

\item{duration_max}{numeric, maximum duration of single signal}
}
\description{
Creates directories and a R list with all the needed tRackIT information for one animal. R list is stored as .rds in animals directory-projroot/data/individuals/animalID/animalID_idFile.rds
}
\author{
Marvin Ludwig & Jannis Gottwald
}
