% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_match_logger_tRackIT.R
\name{time_match_logger_tRackIT}
\alias{time_match_logger_tRackIT}
\title{time match logger data}
\usage{
time_match_logger_tRackIT(animal, path_to_data)
}
\arguments{
\item{path_to_data}{string, path to filtered data}

\item{animalID}{string, label of the tagged individual}
}
\description{
timematch of simoultaneously received signals by logger of one station
}
\author{
Jannis Gottwald
}
