% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRackIT_fun.R
\name{tRackIT}
\alias{tRackIT}
\title{calculate triangulations}
\usage{
tRackIT(animal, projList, tw = 5, be_col = "bearings_filtered")
}
\arguments{
\item{animal}{list, list generated by initAnimal functiom}

\item{projList}{list, list generated by initProject functiom}

\item{tw}{numeric, time window in which bearings are used for triangulation}

\item{be_col}{character, colname of bearing column}
}
\description{
calculates triangulations
}
\author{
Jannis Gottwald
}
